Improve CLI execution error flow
